+++ src/components/OnboardingBanner/ActionCompletedRow.tsx
