+++ src/components/OnboardingBanner/CTARow.tsx
@@ -1,6 +1,6 @@
 import React, { FC, ReactNode } from 'react';
 import { PlusCircleIcon } from '@heroicons/react/16/solid';
-import { ActionCompletedRow } from '@/components/OnboardingBanner/components/ActionCompletedRow';
+import { ActionCompletedRow } from '@/components/OnboardingBanner/ActionCompletedRow';
 
 interface CTARowProps {
   text: string;
